DEMRUM-2027: refactoring dsym upload for modularity
